+++ components/Sponsors-2020.js
@@ -2,7 +2,7 @@ import Link from 'next/link';
 import LazyLoad from 'react-lazyload';
 import { shuffle } from './Shuffler';
 
-const sponsorList = require('../lists/sponsors-2019').sponsors;
+const sponsorList = require('../lists/sponsors-2020').sponsors;
 
 export const Sponsors = () => {
   const WhalePackage = sponsorList.whalePkg.map(s => (
@@ -86,6 +86,28 @@ export const Sponsors = () => {
   return (
     <div className="row center-xs sponsors">
       <div className="col-xs-11 pt-small pb-large bb">
+
+        {/* <div className="row center-xs middle-xs col-xs-10 col-xs-offset-1 border__lines">
+
+          <div className="col-xs-12 center-xs col-sm-4 start-sm">
+            <p style={{ padding: '10px 0' }}>
+              Executive
+              <br className="hidden-sm" />
+              {' '}
+              Sponsor
+            </p>
+          </div>
+
+          <div className="col-xs-10 col-sm-7">
+            <div className="sponsors__single" role="presentation">
+              <a href="mailto:contact@vergecurrency.com?subject=Third Verge Meetup - Sponsorship" target="_blank" rel="noopener noreferrer">
+                <LazyLoad height="100">
+                  <img className="img-responsive sponsors__logo" src="../static/img/meetup-2020/sponsors/new.jpg" style={{ maxHeight: '100px' }} alt="You?" />
+                </LazyLoad>
+              </a>
+            </div>
+          </div>
+        </div> */}
 
         <div className="row start-xs middle-xs pt">
           <div className="col-xs-12 col-sm-3 col-lg-2 end-sm">
