+++ lists/tgadmins.json
@@ -6,39 +6,14 @@
       "twitter": "https://twitter.com/DrogertCrypto"
     },
     {
-      "name": "Lilli",
-      "telegram": "https://t.me/StarLawl",
-      "twitter": "https://twitter.com/mrsFagerdal"
-    },
-    {
-      "name": "Gisn",
-      "telegram": "https://t.me/GISN81",
-      "twitter": "https://twitter.com/Gisn1"
-    },
-    {
-      "name": "Karan Patel",
-      "telegram": "https://t.me/Karan8798",
-      "twitter": "https://twitter.com/Karan_patell1"
-    },
-    {
       "name": "LewLew",
       "telegram": "https://t.me/VergeSlut",
       "twitter": "https://twitter.com/VergeSlut"
     },
     {
-      "name": "Andrew Maan",
-      "telegram": "https://t.me/Andrewmaan",
-      "twitter": "https://twitter.com/andrewhierden"
-    },
-    {
       "name": "Hum",
       "telegram": "https://t.me/HULSE",
       "twitter": "https://twitter.com/HUMXVG"
-    },
-    {
-      "name": "Grandpriest",
-      "telegram": "https://t.me/grandpriest",
-      "twitter": "https://twitter.com/cryptooza"
     },
     {
       "name": "Wolfie",
@@ -79,11 +54,6 @@
       "name": "m3L1sS4",
       "telegram": "https://t.me/xMithras",
       "twitter": "https://twitter.com/ecilaam"
-    },
-    {
-      "name": "Rbull",
-      "telegram": "https://t.me/Rbull076",
-      "twitter": "https://twitter.com/Rbull076"
     },
     {
       "name": "Vergenia",
