+++ next.config.js
@@ -45,10 +45,11 @@ module.exports = {
       '/wallets/terms': { page: '/wallets/terms' },
       '/wallets': { page: '/wallets' },
       '/donationrewards': { page: '/donationrewards' },
-    //  '/get-started': { page: '/get-started' },
+      //'/get-started': { page: '/get-started' },
       '/getstarted': { page: '/getstarted' },
       '/pressreleases': { page: '/pressreleases' },
       '/meetup-2018': { page: '/meetup-2018' },
+      '/meetup-2019': { page: '/meetup-2019' },
       '/meetup': { page: '/meetup' },
       '/partnership-donation-rewards': { page: '/partnership-donation-rewards' },
       '/partnership-donation-mentions': { page: '/partnership-donation-mentions' },
