+++ package.json
@@ -36,41 +36,40 @@
     "babel-plugin-inline-react-svg": "^0.5.2",
     "babel-plugin-module-resolver": "^2.7.1",
     "babel-plugin-wrap-in-js": "^1.1.0",
-    "compression": "^1.7.2",
-    "cryptowatch-embed": "^0.3.2",
+    "compression": "^1.7.4",
     "css-loader": "^0.28.9",
     "deep-filter": "^1.0.2",
-    "express": "^4.16.2",
+    "express": "^4.17.1",
     "gh-pages": "^1.1.0",
-    "glob": "^7.1.2",
+    "glob": "^7.1.4",
     "glob-all": "^3.1.0",
     "gulp": "^3.9.1",
     "gulp-webserver": "^0.9.1",
     "i18next": "^10.3.0",
     "i18next-browser-languagedetector": "^2.1.0",
-    "i18next-express-middleware": "^1.0.10",
+    "i18next-express-middleware": "^1.8.2",
     "i18next-node-fs-backend": "^1.0.0",
     "i18next-xhr-backend": "^1.5.1",
     "isomorphic-unfetch": "^2.0.0",
     "markdown-in-js": "^1.1.3",
     "moment": "^2.20.1",
-    "moment-timezone": "^0.5.14",
+    "moment-timezone": "^0.5.26",
     "next": "4.2.3",
-    "node-sass": "^4.4.0",
+    "node-sass": "^4.12.0",
     "normalize.css": "^7.0.0",
     "npm-update-all": "^1.0.1",
     "nprogress": "^0.2.0",
     "platform": "^1.3.5",
     "postcss-easy-import": "^3.0.0",
     "postcss-loader": "^2.1.0",
-    "prop-types": "^15.6.0",
+    "prop-types": "^15.7.2",
     "raw-loader": "^0.5.1",
-    "react": "^16.2.0",
-    "react-burger-menu": "^2.2.3",
-    "react-dom": "^16.2.0",
+    "react": "^16.9.0",
+    "react-burger-menu": "^2.6.11",
+    "react-dom": "^16.9.0",
     "react-expand-collapse": "^0.2.0",
     "react-i18next": "^7.3.4",
-    "react-lazyload": "^2.3.0",
+    "react-lazyload": "^2.6.2",
     "react-moment": "^0.7.0",
     "react-web-tabs": "^0.4.3",
     "sass-loader": "^6.0.6",
@@ -80,15 +79,15 @@
     "webpack": "^3.10.0"
   },
   "devDependencies": {
-    "eslint": "^4.18.0",
-    "eslint-config-airbnb": "^16.1.0",
+    "eslint": "5.3.0",
+    "eslint-config-airbnb": "17.1.0",
     "eslint-config-airbnb-standard": "^1.6.6",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.7.0",
-    "sass-lint": "^1.12.1",
+    "eslint-plugin-import": "2.14.0",
+    "eslint-plugin-jsx-a11y": "6.1.1",
+    "eslint-plugin-react": "7.11.0",
+    "sass-lint": "^1.13.1",
     "sw-precache-webpack-plugin": "^0.11.5",
-    "uglify-js": "https://github.com/mishoo/UglifyJS2#harmony-v2.8.22",
+    "uglify-js": "git+https://github.com/mishoo/UglifyJS2.git#harmony-v2.8.22",
     "webpack-bundle-analyzer": ">=3.3.2"
   }
 }
